Moved method `WaitForConsensusReadyV2` from `Consensus` struct. (#4803)
